Fix bind group dynamic offsets
setBindGroup has two overloads when using dynamic offsets. One taking a sequence<GPUBufferDynamicOffset> and one taking a Uint32Array with an offset and size. If the data was a Uint32Array, we assumed the second overload was used, however a Uint32Array can be implicitly converted to a sequence<GPUBufferDynamicOffset>. Fix this by also checking the presence of the offset and size arguments.